float

+++ src/eval/patch_predict.py
@@ -323,8 +323,8 @@ def evaluate_seg(data_loader, finetuned_encoder, device, identifier, patch_size_
             # check that all predictions are between 0 and 1
             assert logits.min() >= 0 and logits.max() <= 1
 
-            all_preds_1D.append(rearrange(torch.squeeze(logits), "b s -> (b s)").cpu().numpy())
-            all_labels_1D.append(rearrange(labels, "b h w -> (b h w)").cpu().numpy())
+            all_preds_1D.append(rearrange(torch.squeeze(logits), "b s -> (b s)").float().cpu().numpy())
+            all_labels_1D.append(rearrange(labels, "b h w -> (b h w)").float().cpu().numpy())
 
             spatial_patches_per_dim = int(logits.shape[1] ** 0.5)
             logits = rearrange(
